requirements.txt
# Backend dependencies
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3

# Data processing
pandas==2.1.4
openpyxl==3.1.2

# Database
sqlalchemy==2.0.25

# AI
anthropic==0.18.1

# Report generation
jinja2==3.1.3
# weasyprint==62.0  # PDF generation (optional, requires system dependencies)

# Development
python-dotenv==1.0.0
